requirements.txt
Flask==2.3.3
azure-monitor-opentelemetry==1.0.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-logging==0.41b0
python-dotenv==1.0.0